requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
opencv-python==4.5.3.56
numpy==1.21.2
scikit-image==0.18.3
scikit-learn==0.24.2
xgboost==1.4.2
werkzeug==2.0.1
pyopenssl==20.0.1
gunicorn==20.1.0 